120 bpm / 140 bpm clicktracks + more
i added new clicktracks because the one provided was NOT 120 bpm. in doing so, i created the appropriate banks, which is how Unity connects to FMOD.

The main camera is a listener but the object is attenuated from the Player because technically the player is viewing the game from the camera but will listen from the player's perspective.

The Player has an Event Emitter that starts the clicktrack on "Object Started" and "Object Destroyed" which can be changed to different triggers that work for a given object. This will most likely change.